Make prometheus configurable at compile time
